Add metric for tracking Bigtable block uploads
# What

This commit adds a datapoint metric called
"storage-bigtable-query-last-block-uploaded", which communicates the
most recently uploaded block by the threads spawned inside the
"LedgerStorage::upload_confirmed_block" method.

# Why

It gets the information into a time series database-compatible format,
which makes it a easier to build monitoring and alerting.

# Testing

We built the binary locally using the solanalabs/rust container image,
and tested the binary on a machine and saw log messages that looked like
this:

```
[2022-06-17T14:01:43.843999910Z INFO  solana_metrics::metrics] datapoint: storage-bigtable-query-last-block-uploaded height=135987534i
[2022-06-17T14:01:43.871708319Z INFO  solana_metrics::metrics] datapoint: storage-bigtable-query-last-block-uploaded height=135987530i
[2022-06-17T14:01:43.875518939Z INFO  solana_metrics::metrics] datapoint: storage-bigtable-query-last-block-uploaded height=135987526i
[2022-06-17T14:01:43.985200230Z INFO  solana_metrics::metrics] datapoint: storage-bigtable-query-last-block-uploaded height=135987520i
[2022-06-17T14:01:44.030815720Z INFO  solana_metrics::metrics] datapoint: storage-bigtable-query-last-block-uploaded height=135987528i
[2022-06-17T14:01:44.126093176Z INFO  solana_metrics::metrics] datapoint: storage-bigtable-query-last-block-uploaded height=135987524i
```